--- app/http/handler/post/index.go
@@ -1,24 +1,0 @@
-package post
-
-import (
-	rePost "golang-odai/app/http/render/post"
-	"golang-odai/domain/post"
-	"net/http"
-)
-
-type Data struct{
-	Posts []post.Post
-}
-
-func (p *Post) Index(w http.ResponseWriter, r *http.Request) {
-	posts, err := p.repoPost.FindAll(r.Context())
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-	}
-
-	if err := p.re.Index(w, rePost.IndexData{
-		Posts: posts,
-	}); err != nil {
-		panic(err)
-	}
-}
