--- app/http/handler/signup/user.go
@@ -1,19 +1,0 @@
-package signup
-
-import (
-	"github.com/unrolled/render"
-	"golang-odai/app/http/render/signup"
-	"golang-odai/domain/user/repository"
-)
-
-type User struct {
-	re *signup.Render
-	repoUser repository.User
-}
-
-func New(re *render.Render, p repository.User) *User {
-	return &User{
-		re:       signup.New(re),
-		repoUser: p,
-	}
-}
