+++ domain/post.go
@@ -1,0 +1,9 @@
+package domain
+
+type Post struct {
+	ID uint32
+	UserID uint32
+	Text string
+
+	User *User
+}
