--- domain/post/factory/posts.go
@@ -1,1 +1,0 @@
-package factory
