+++ domain/user.go
@@ -1,0 +1,8 @@
+package domain
+
+type User struct {
+	ID uint32
+	Name string
+	Email string
+	Password string
+}
