--- domain/user/user.go
@@ -1,8 +1,0 @@
-package user
-
-type User struct {
-	ID uint32
-	Name string
-	Email string
-	Password string
-}
